Bug in using argparse and multiple repetition of default values
Using the script I noticed that there was a bug in the commit automation script checking. The parameters had been reversed. Also, the commit flag, so to speak, the -g had wrong parameters that forced it to have an argument.

Furthermore, as explained in the issue, there is a need to change the readme, as it is incorrectly recommended to use the script. I recommend the following change: python yt_crawl.py -a "API_KEY_HERE" -g